fix cycle in the dependency
